+++ src/arkreactor/Compiler/Macros/Executors/Symbol.cpp
@@ -10,14 +10,14 @@ namespace Ark::internal
     bool SymbolExecutor::applyMacro(Node& node)
     {
         // error ?
-        Node* macro = findNearestMacro(node.string());
+        const Node* macro = findNearestMacro(node.string());
 
         if (macro != nullptr)
         {
             // !{name value}
             if (macro->constList().size() == 2)
             {
-                node = macro->list()[1];
+                node = macro->constList()[1];
                 return true;
             }
         }